+++ quasar/src/qs/vertex/IndexBuffer.cpp
@@ -21,7 +21,7 @@ namespace qs
 
 	void IndexBuffer::create(const std::vector<unsigned int>& indexs, const qs::BufferType bufferType) noexcept
 	{
-		m_count = indexs.size();
+		m_count = static_cast<unsigned int>(indexs.size());
 
 		glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, m_id);
